// src/app/core/models/rolePrivileges.ts
export enum Privileges {
  USER_LISTING = 'USER_LISTING',
  SELLER_LISTING = 'SELLER_LISTING',
  CONTROL_PANEL = 'CONTROL_PANEL',
  CATEGORY_LISTING = 'CATEGORY_LISTING',
  PHOTOGRAPHER_MANAGEMENT = 'PHOTOGRAPHER_MANAGEMENT',
  AUCTION_MANAGEMENT = 'AUCTION_MANAGEMENT',
  CONTENT_MANAGEMENT = 'CONTENT_MANAGEMENT',
}

export enum Roles {
  SUPER_ADMIN = 'superadmin',
  ADMIN = 'admin',
}